Reduce endpoints. Embed user input in "obj" field.
The endpoints available are:
/{namespace}/systems/{systemId}
/{namespace}/systems/{systemId}/agents/{agentId}
/{namespace}/systems/{systemId}/agents/{agentId}/jvms/{jvmId}

The inputs are now embedded:
{
  "obj" { <valid json input> },
  "systemId",
  ...
}

The input identities are attached outside of the embedded object. This
makes it easier to return the input given in a PUT request in
subsequent GET/POST requests.

The input parameters for Ids accept "all" for GET/POST/DELETE requests
but not for PUT requests.